src/App.js: refactor getData
callDateRange now determines what data is requested from the server
based on the previous set value for callDateRange and the most recently
received call.

The autoload side effect now simply sets the callDateRange after the
specified autoloadDelay and relies on the updateRange side effect to query
data for the new range.